app/backend/requirements.txt
fastapi==0.121.1
uvicorn==0.30.0
passlib==1.7.4
bcrypt==4.1.2
pydantic==2.12.4
pydantic_settings==2.11.0
python_jose==3.5.0
SQLAlchemy==2.0.44
alembic==1.13.1
psycopg2-binary==2.9.10
python-multipart==0.0.20
black==25.11.0
python-dotenv==1.0.0
